Also plot the secondary filament count
It would be nice if we 'hue' the filament count with the fish number,
but for this we need to thoroughly massage the data, since Matthias' XLS
file is not suited to import into the dataframe in its current state.
We would have to change it considerably...
Nonetheless, this fixes #2 for the moment.