I should get a massage therapist license :)
We're <del>coercing</del> massaging the input XLS files into nice
dataframes so we can easily 'hue' the plots later on.
Also moving one cell up from the bottom to the top, so all the data
input is a the beginning of the notebook, then we have some image
processing parts, then we have the output part.