Massage filament data, Plot more BRDU
Reviewer2 would like to have a regression analysis of filament growth
and mitotic index.
For this we need to massage the filament data into a proper dataframe so
we can seaborn.lmplot it.
Also prepare Fig07 for the new images that Helena will provide soon by
rearranging it into a new layout and preparing 4x4 subplots on the right
side of it while plotting the mitotic index on the top/tips and
bottom/bases to the left.