Add raw data of the new microscopy images
Reviewer1 asked for clarification of Fig07.
Helena provided new microscopy images (for the median value of both at
the tips and bases of the gills for both controls and swimmers).
We only add the (original) TIFF files to the repository, the single
frames can easily be generated with ImageJ, as specified in the README
of this folder.
The single PNG frames are `.gitignore`d